Move autogen sample to separate project
