--- samples/02-autogen/Plugins/OutputFileInfo.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace _02_autogen.Plugins
-{
-    internal class OutputFileInfo
-    {
-        public string FileName { get; set; }
-
-        public long Size { get; set; }
-
-        public string Path { get; set; }
-    }
-}
