+++ src/SemanticKernel.Assistants.AutoGen/Assistants/AssistantAgent.yaml
