+++ src/SemanticKernel.Assistants.AutoGen/Assistants/CodeInterpreter.yaml
