+++ src/SemanticKernel.Assistants.AutoGen/Plugins/OutputFileInfo.cs
@@ -1,0 +1,12 @@
+﻿// Copyright (c) Kevin BEAUGRAND. All rights reserved.
+
+namespace SemanticKernel.Assistants.AutoGen.Plugins;
+
+internal class OutputFileInfo
+{
+    public string FileName { get; set; }
+
+    public long Size { get; set; }
+
+    public string Path { get; set; }
+}
